+++ lib/cobrato/resources/payment.rb
@@ -21,6 +21,12 @@ module Cobrato
         end
       end
 
+      def reschedule(id, params)
+        http.post("#{resource_base_path}/#{id}/reschedule", params: params) do |response|
+          respond_with_entity response
+        end
+      end
+
       def register_error(id)
         http.post("#{resource_base_path}/#{id}/register_error") do |response|
           respond_with_entity response